requirements.txt
aiogram
playwright
python-dotenv
nest_asyncio
aiohttp
apscheduler
